requirements.txt
SQLAlchemy==1.3.22
flask-sqlalchemy==2.4.4
pymysql==0.10.1
flask-bcrypt==0.7.1
pyjwt==1.4.2
